整理auth模块的接口

+++ roses-gateway/src/main/java/com/stylefeng/roses/gate/filter/PathMatchFilter.java
@@ -58,7 +58,7 @@ public class PathMatchFilter extends ZuulFilter {
         HttpServletResponse response = currentContext.getResponse();
 
         String userId = UserIdHolder.get();
-        Set<String> permissionUrls = authServiceConsumer.getUserPermissionUrls(Integer.valueOf(userId));
+        Set<String> permissionUrls = authServiceConsumer.getUserPermissionUrls(Long.valueOf(userId));
 
         AntPathMatcher antPathMatcher = new AntPathMatcher();
         boolean hasPermission = false;
